test(bulk-model-sync): reuse common test cases
